refactor: 💡 Added getBaseSchema function
