+++ .vscode/settings.json
@@ -17,6 +17,7 @@
   "Anand",
   "convoform",
   "Gantt",
+  "paralleldrive",
   "treeshake",
   "trpc",
   "Utkarsh"
