+++ packages/db/src/schema/conversations/conversation.ts
@@ -4,6 +4,7 @@ import { boolean, jsonb, pgTable, text, timestamp } from "drizzle-orm/pg-core";
 import { createInsertSchema, createSelectSchema } from "drizzle-zod";
 import { z } from "zod";
 
+import { getBaseSchema } from "../base";
 import { form } from "../forms/form";
 import { CollectedData, Transcript } from "./validation";
 
@@ -14,12 +15,7 @@ export const fieldHavingDataSchema = z.object({
 export type FieldHavingData = z.infer<typeof fieldHavingDataSchema>;
 
 export const conversation = pgTable("Conversation", {
-  id: text("id")
-    .primaryKey()
-    .$defaultFn(() => createId())
-    .unique(),
-  createdAt: timestamp("createdAt").defaultNow().notNull(),
-  updatedAt: timestamp("updatedAt").defaultNow().notNull(),
+  ...getBaseSchema(),
   name: text("name").notNull(),
   transcript: jsonb("transcript").array().$type<Transcript[]>(),
   collectedData: jsonb("collectedData")
