+++ packages/db/src/schema/formFields/formField.ts
@@ -4,15 +4,11 @@ import { pgTable, text, timestamp } from "drizzle-orm/pg-core";
 import { createInsertSchema, createSelectSchema } from "drizzle-zod";
 import { z } from "zod";
 
+import { getBaseSchema } from "../base";
 import { form } from "../forms/form";
 
 export const formField = pgTable("FormField", {
-  id: text("id")
-    .primaryKey()
-    .$defaultFn(() => createId())
-    .unique(),
-  createdAt: timestamp("createdAt").defaultNow().notNull(),
-  updatedAt: timestamp("updatedAt").defaultNow().notNull(),
+  ...getBaseSchema(),
   // Human readable name of the field, also will be display in table column
   fieldName: text("fieldName").notNull(),
   // Used while generating question for the field
